% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amateuR.R
\docType{package}
\name{amateuR}
\alias{amateuR}
\alias{amateuR-package}
\title{Prediction for amateur German football outcomes}
\description{
amateuR is an R Package for obtaining data for amateur German football leagues and predicting future matches with a Dixon Coles Model.
}
\details{
The package provides functions to download German amateur football data from kicker.de by using web scraping methods and the rvest package.
Based on the downloaded data the table, individual game results and the date and time of future games can be shown.
It also provides a function for predicting future games.
Some of the results can visualized using specific plot functions.
}
\author{
James Travis <james.travis@stud.uni-goettingen.de>

Niklas Hagemann <niklas.hagemann@stud.uni-goettingen.de>
}
