% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{data}
\alias{data}
\title{data}
\format{An object of class \code{character} of length 1935.}
\usage{
data
}
\description{
Dataset for testing the amateuR package.
The Dataset contains the data of the Kreisliga Göttingen-Osterode of the season 2019-20, which was downloaded at 2019-08-23.
}
\keyword{datasets}
