% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/URL_and_download.R
\name{download_league_data}
\alias{download_league_data}
\title{Data download from kicker.de}
\usage{
download_league_data(league_url)
}
\arguments{
\item{league_url}{The URL of the kicker.de webpage of the league of interest. Please create this with \code{\link[amateuR]{build_league_url}}. Alternativly you can copy & paste it manually, but make sure it ends with \code{"/-1"}.}
}
\description{
This function downloads data from kicker.de! Please note that this function is only for the download of German amateur football leagues. It works mostly for professional and international leagues, but occasionally errors can occur here.
}
\examples{
download_league_data(
  "https://www.kicker.de/berlin-kreisklasse-a-staffel-1-6322/spieltag/2018-19/-1"
)

download_league_data(build_league_url(
  "berlin-kreisklasse-a-staffel-1-6322",
  season = "2018-19"
))
}
