% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tau_and_initialise.R
\name{estimate_params}
\alias{estimate_params}
\title{Initialising data}
\usage{
estimate_params(data, alpha = 0.9)
}
\arguments{
\item{data}{Output of \code{\link[amateuR]{all_game_data}} without specifying a team}

\item{alpha}{Significance level for confidence intervals}
}
\description{
Set up and optimise a bipoisson likelihood function with the DC correction.
This function returns parameter estimates with Wald Confidence intervals.
}
\examples{
my_games <- all_game_data(download_league_data(
  "https://www.kicker.de/berlin-kreisklasse-a-staffel-1-6322/spieltag/2018-19/-1"
))

estimate_params(data = my_games, alpha = 0.9)
}
