% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_plots.R
\name{plot_goal_difference}
\alias{plot_goal_difference}
\title{Plot the goal differences}
\usage{
plot_goal_difference(game_data, team, cumulated = TRUE)
}
\arguments{
\item{game_data}{Output of the function \code{\link[amateuR]{all_game_data}}.}

\item{team}{A team name as string.}

\item{cumulated}{If \code{TRUE}, the cummulated goal differences are plotted, if \code{FALSE} the individual differences for every game are plotted.}
}
\description{
This function plots the goal differences for one specified team.
}
\details{
Description of the colours of the data points:
\tabular{ll}{
Green \tab  Matches won \cr
Orange \tab  Matches drawn \cr
Red \tab  Matches lost
}
}
\examples{
data <- all_game_data(amateuR::kreisliga_goettingen)
plot_goal_difference(game_data = data, team = "RSV Göttingen 05", cumulated = TRUE)
plot_goal_difference(game_data = data, team = "RSV Göttingen 05", cumulated = FALSE)
}
