% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_plots.R
\name{plot_points}
\alias{plot_points}
\title{Plot the points}
\usage{
plot_points(game_data, team)
}
\arguments{
\item{game_data}{Output of the function \code{\link[amateuR]{all_game_data}}.}

\item{team}{A team name as string.}
}
\description{
This function plots the cummulated points for one specified team.
}
\details{
Description of the colours of the data points:
\tabular{ll}{
Green \tab  Matches won \cr
Orange \tab Matches drawn \cr
Red \tab  Matches lost
}
}
\examples{
data <- download_league_data("https://www.kicker.de/kreisliga-osterode-fb-1/spieltag/2019-20/-1")
data <- all_game_data(data)
plot_points(game_data = data, team = "RSV Göttingen 05")
}
