% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_plots.R
\name{plot_single_match}
\alias{plot_single_match}
\title{Plotting the Dixon Coles prediction with a confidence interval for a given match}
\usage{
plot_single_match(game_data, estimate, home, away)
}
\arguments{
\item{game_data}{Output of \code{\link[amateuR]{all_game_data}} with parameter \code{team = NA}}

\item{estimate}{Output of \code{\link[amateuR]{estimate_params}} using the same data}

\item{home}{Home team name as a string}

\item{away}{Away team name as a string}
}
\description{
This function allows the user to plot the Dixon Coles estimate in a 2d diagram with the home team on the x axis and the away team on the y axis.
}
\details{
\tabular{ll}{
Red point \tab The final full time score for a past match\cr
Black circle \tab The point estimate \cr
Black box \tab  Confidence interval
}
}
\examples{
my_games <- all_game_data(amateuR::kreisliga_goettingen)
estimate <- estimate_params(data = my_games, alpha = 0.9)

plot_single_match(
  game_data = my_games, estimate = estimate,
  home = "RSV Göttingen 05", away = "TSC Dorste"
)
}
