% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions.R
\name{prediction_interval}
\alias{prediction_interval}
\title{Point estimate fuction and confidence intervals for footbal predictions}
\usage{
prediction_interval(data, estimate, teamname, method = "future")
}
\arguments{
\item{data}{Output of \code{\link[amateuR]{all_game_data}} with parameter \code{team = NA}}

\item{estimate}{Output of \code{\link[amateuR]{estimate_params}} using the same data}

\item{teamname}{The name of the team of intrest as string}

\item{method}{Optional parameter which takes the values \code{"all"}, \code{"past"} or \code{"future"} to filter predictions}
}
\description{
This function returns point estimates for the number of home and away goals for all remaining matches for one team.
}
\examples{
my_games <- all_game_data(download_league_data(
  "https://www.kicker.de/berlin-kreisklasse-a-staffel-1-6322/spieltag/2018-19/-1"
))
estimate <- estimate_params(data = my_games, alpha = 0.9)
#
prediction_interval(
  data = my_games, estimate = estimate,
  teamname = "1. FC Schöneberg III", method = "all"
)
}
