% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_data.R
\name{started_season_data_wrangling}
\alias{started_season_data_wrangling}
\title{started_season_data_wrangling}
\usage{
started_season_data_wrangling(data, future_games)
}
\description{
This is an internal function for all_game_data(), it performs the data wrangling if the season is already started.
}
\keyword{internal}
