% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_data.R
\name{sv}
\alias{sv}
\title{Shift a vector}
\usage{
sv(vec, shift, sorted = TRUE)
}
\arguments{
\item{v}{A numerical vector}

\item{l}{A numerical vector with the shifting parameters}
}
\description{
This is an internal function for all_game_data(), it shifts the values in vec by the values of the shift vector. The new vector is sorted.
}
\keyword{internal}
