% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tau_and_initialise.R
\name{tau}
\alias{tau}
\title{Tau(x,y)}
\usage{
tau(x, y, lambda, mu, rho)
}
\arguments{
\item{x}{number of home goals}

\item{y}{number of away goals}

\item{lambda}{expected number of home goals}

\item{mu}{expected number of away goals}

\item{rho}{correlation term}
}
\description{
This is an internal function for \code{\link[amateuR]{negloglik}}. It allows you to compensate for the lack of independence in the bivariate Poisson Model.
}
\keyword{internal}
