% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_data.R
\name{all_game_data}
\alias{all_game_data}
\title{Extract past gameday results and date and time of future games from the league data vector}
\usage{
all_game_data(league_data, team = NA)
}
\arguments{
\item{league_data}{A league data vector, which is created with \code{\link[amateuR]{download_league_data}}.}

\item{team}{\code{NA} or an optional parameter to filter the list for a given team name}
}
\description{
This function extracts past gameday results and date and time of future games from the league data vector, which is created with \code{\link[amateuR]{download_league_data}}.
}
\examples{
all_game_data(download_league_data(
  "https://www.kicker.de/berlin-kreisklasse-a-staffel-1-6322/spieltag/2018-19/-1"
))
}
