% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_data.R
\name{all_game_data_filter}
\alias{all_game_data_filter}
\title{all_game_data_filter}
\usage{
all_game_data_filter(L, team)
}
\arguments{
\item{L}{Unfiltered result}

\item{team}{NA, a team name or a vector of team names. If NA (default), all games are returned. If team is specified, only the games of this team(s) are returned. If you want to use the result for prediction, use NA.}
}
\description{
This is an internal function for all_game_data(), it perfoms the optional filter for all_game_data().
}
\keyword{internal}
