% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/current_table.R
\name{current_table}
\alias{current_table}
\title{Extract the current table from the league data vector}
\usage{
current_table(league_data)
}
\arguments{
\item{league_data}{A league data vector, which is created with \code{\link[amateuR]{download_league_data}}.}
}
\description{
This function extracts the current table from the league data vector, which is created with \code{\link[amateuR]{download_league_data}}.
}
\examples{
current_table(download_league_data(
  "https://www.kicker.de/berlin-kreisklasse-a-staffel-1-6322/spieltag/2018-19/-1"
))
}
