% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tau_and_initialise.R
\name{estimate_params}
\alias{estimate_params}
\title{Initialising data}
\usage{
estimate_params(data, alpha = 0.9)
}
\arguments{
\item{data}{Output of \code{\link[amateuR]{all_game_data}} without specifying a team}

\item{alpha}{Significance level for confidence intervals}
}
\description{
This function sets up and optimises a bipoisson likelihood function with the DC correction. It returns parameter estimates with Wald Confidence intervals.
}
\examples{
my_games <- all_game_data(amateuR::kreisliga_goettingen)
estimate_params(data = my_games, alpha = 0.9)
}
