% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tau_and_initialise.R
\name{negloglik}
\alias{negloglik}
\title{Negative loglikelihood function}
\usage{
negloglik(params, goals_home, goals_visitor, team_home, team_visitor,
  structure)
}
\arguments{
\item{params}{likelihood param}

\item{goals_home}{likelihood param}

\item{goals_visitor}{likelihood param}

\item{team_home}{likelihood param}

\item{team_visitor}{likelihood param}

\item{structure}{structure}
}
\description{
This is an internal function for estimates()
}
\keyword{internal}
