% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_plots.R
\name{plot_predictions}
\alias{plot_predictions}
\title{Plotting the Dixon Coles prediction with confidence interval for a single team}
\usage{
plot_predictions(game_data, estimate, team_name, method = "future")
}
\arguments{
\item{game_data}{Output of amateuR::all_game_data() with parameter team = NA}

\item{estimate}{Output of amateuR::estimate_params() using the same data}

\item{team_name}{Home team name as string}

\item{method}{method = "futureH" returns all predictions for future home matches \cr
method = "futureHA" returns all predictions for future home and away matchs \cr
method = "allH" returns all predictions for past and future home matches}
}
\description{
This function allows the user to plot the Dixon Coles estimate in a 2d diagram for a given team.
Use the optional argument "method" to return all or just future matches.
}
\examples{
my_games <- all_game_data(download_league_data(
  "https://www.kicker.de/berlin-kreisklasse-a-staffel-1-6322/spieltag/2018-19/-1"
))
estimate <- estimate_params(data = my_games, alpha = 0.9)
plot_predictions(game_data = my_games, estimate = estimate, "VfB Hermsdorf III", method = "allH")
}
