% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_plots.R
\name{plot_single_match}
\alias{plot_single_match}
\title{Plotting the Dixon Coles prediction with confidence interval for a given match}
\usage{
plot_single_match(game_data, estimate, home, away)
}
\arguments{
\item{game_data}{Output of amateuR::all_game_data() with parameter team = NA}

\item{estimate}{Output of amateuR::estimate_params() using the same data}

\item{home}{Home team name as string}

\item{away}{away team name as string}
}
\description{
This function allows the user to plot the dixon Coles estimate in a 2d diagram with the home team on the x axis and the away team on the y axis
}
\examples{
my_games <- all_game_data(download_league_data(
  "https://www.kicker.de/berlin-kreisklasse-a-staffel-1-6322/spieltag/2018-19/-1"
))
estimate <- estimate_params(data = my_games, alpha = 0.9)

plot_single_match(
  game_data = my_games, estimate = estimate,
  home = "1. FC Schöneberg III", away = "Berliner SC III"
)
}
