% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions.R
\name{prediction}
\alias{prediction}
\title{Point estimate function for football predictions}
\usage{
prediction(data, estimate, teamname, method = "future")
}
\arguments{
\item{data}{Output of \code{\link[amateuR]{all_game_data}} with parameter \code{team = NA}}

\item{estimate}{Output of \code{\link[amateuR]{estimate_params}} using the same data}

\item{teamname}{The name of the team of interest as a string}

\item{method}{Optional parameter which takes the values \code{"all"}, \code{"past"} or \code{"future"} to filter predictions}
}
\value{
A description of the columns of both prediction data frames, upcoming and future:
\tabular{ll}{
 \code{Comment} \tab A simple warning if the prediction may not be accurate \cr
 \code{Home Estimate} \tab Prediction of the number of goals scored by the \emph{home team}, rounded to nearest integer\cr
 \code{Away Estimate} \tab Prediction of the number of goals scored by the \emph{away team}, rounded to nearest integer \cr
 \code{Final Score} \tab The full time score for past games
}
}
\description{
This function returns point estimates for the number of home and away goals for all remaining matches for one team.
}
\examples{
my_games <- all_game_data(download_league_data(
  "https://www.kicker.de/berlin-kreisklasse-a-staffel-1-6322/spieltag/2018-19/-1"
))
estimate <- estimate_params(data = my_games, alpha = 0.9)
prediction(data = my_games, estimate = estimate, teamname = "1. FC Schöneberg III", method = "all")
}
