% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_data.R
\name{sv}
\alias{sv}
\title{Shift a vector}
\usage{
sv(vec, shift, sorted = TRUE)
}
\arguments{
\item{sorted}{If \code{TRUE}, the returned vector is sorted}

\item{v}{A numerical vector to shift}

\item{l}{A numerical vector with the shifting parameters}
}
\description{
This is an internal function for \code{\link[amateuR]{all_game_data}}, it shifts the values in vec by the values of the shift vector.
}
\keyword{internal}
